correction float int dans addEdges et ajout rapport

+++ include/graphe.h
@@ -17,7 +17,7 @@ struct Graph {
 // Prototypes des fonctions
 struct Graph* createGraph(int vertices, int isDirected);
 int edgeExists(const struct Graph* graph, int src, int dest);
-void addEdge(struct Graph* graph, int src, int dest, int weight);
+void addEdge(struct Graph* graph, int src, int dest, float weight);
 // void printGraph(struct Graph* graph); // pour imprimer la liste d'adjacence
 int degreSortant(const struct Graph* graph, int station);
 void freeGraph(struct Graph* graph);
